pipeline/pipeline_requirements.txt
pandas
pymysql
python-dotenv
pyarrow
awswrangler
